commit a5293cb58082f5a9044b9f335ebd001283eb4e43 from orca-snoplus
Original commit message:
add config file testing

+++ Source/Objects/Misc Objects/CouchDB/ORCouchDBModel.m
@@ -84,8 +84,6 @@ static NSString* ORCouchDBModelInConnector 	= @"ORCouchDBModelInConnector";
 - (void) periodicCompact;
 - (void) updateDataSets;
 - (void) updateStatus;
-- (void) _cancelAllPeriodicOperations;
-- (void) _startAllPeriodicOperations;
 @end
 
 @implementation ORCouchDBModel
@@ -119,7 +117,11 @@ static NSString* ORCouchDBModelInConnector 	= @"ORCouchDBModelInConnector";
 {
     if(![self aWake]){
         [self createDatabase];
-        [self _startAllPeriodicOperations];
+		[self performSelector:@selector(updateMachineRecord) withObject:nil afterDelay:2];
+		[self performSelector:@selector(updateExperiment) withObject:nil afterDelay:3];
+		[self performSelector:@selector(updateRunInfo) withObject:nil afterDelay:3];
+		[self performSelector:@selector(updateDatabaseStats) withObject:nil afterDelay:4];
+		[self performSelector:@selector(periodicCompact) withObject:nil afterDelay:60];
         [self registerNotificationObservers];
     }
     [super wakeUp];
@@ -128,7 +130,7 @@ static NSString* ORCouchDBModelInConnector 	= @"ORCouchDBModelInConnector";
 
 - (void) sleep
 {
-    [self _cancelAllPeriodicOperations];
+	[NSObject cancelPreviousPerformRequestsWithTarget:self];
     [[NSNotificationCenter defaultCenter] removeObserver:self];
 	[self deleteDatabase];
 	[super sleep];
@@ -201,12 +203,7 @@ static NSString* ORCouchDBModelInConnector 	= @"ORCouchDBModelInConnector";
                      selector : @selector(alarmsChanged:)
                          name : ORAlarmWasPostedNotification
                        object : nil];	
-
-    [notifyCenter addObserver : self
-                     selector : @selector(alarmsChanged:)
-                         name : ORAlarmWasAcknowledgedNotification
-                       object : nil];
-    
+	
     [notifyCenter addObserver : self
                      selector : @selector(alarmsChanged:)
                          name : ORAlarmWasClearedNotification
@@ -289,10 +286,13 @@ static NSString* ORCouchDBModelInConnector 	= @"ORCouchDBModelInConnector";
 - (void) setKeepHistory:(BOOL)aKeepHistory
 {
     [[[self undoManager] prepareWithInvocationTarget:self] setKeepHistory:keepHistory];
-    keepHistory = aKeepHistory;
-    if(keepHistory){
-        [self createHistoryDatabase];
-    }
+	if([self couchRunning]){
+		keepHistory = aKeepHistory;
+		if(keepHistory){
+			[self createHistoryDatabase];
+		}
+	} 
+	else keepHistory=NO;
 
     [[NSNotificationCenter defaultCenter] postNotificationName:ORCouchDBModelKeepHistoryChanged object:self];
 }
@@ -305,17 +305,31 @@ static NSString* ORCouchDBModelInConnector 	= @"ORCouchDBModelInConnector";
 - (void) setStealthMode:(BOOL)aStealthMode
 {
     [[[self undoManager] prepareWithInvocationTarget:self] setStealthMode:stealthMode];
-    stealthMode = aStealthMode;
-    if(stealthMode){
-        if([ORCouchDBQueue operationCount]) [ORCouchDBQueue cancelAllOperations];
-        [self _cancelAllPeriodicOperations];
-        [self deleteDatabase];
-    }
-    else {
-        [self createDatabase];
-        [self _startAllPeriodicOperations];
-    }
+	BOOL okToRun = [self couchRunning];
+	if(okToRun){
+		stealthMode = aStealthMode;
+		if(stealthMode){
+			if([ORCouchDBQueue operationCount]) [ORCouchDBQueue cancelAllOperations];
+			[self deleteDatabase];
+		}
+		else {
+			[self createDatabase];
+		}
+	}
 	[[NSNotificationCenter defaultCenter] postNotificationName:ORCouchDBModelStealthModeChanged object:self];
+}
+
+- (BOOL) couchRunning
+{
+	BOOL okToRun = YES;
+	ORCouchDB* couch = [[[ORCouchDB alloc] init] autorelease];
+	if(![couch couchDBRunning]){
+		NSBeep();
+		NSLogColor([NSColor redColor],@"It appears CouchDB is not running.\n");
+		okToRun = NO;
+	}
+	
+	return okToRun;
 }
 
 - (id) nextObject
@@ -1384,20 +1398,6 @@ static NSString* ORCouchDBModelInConnector 	= @"ORCouchDBModelInConnector";
 		
 		[self performSelector:@selector(updateDataSets) withObject:nil afterDelay:10];
 	}
-}
-
-- (void) _cancelAllPeriodicOperations
-{
-	[NSObject cancelPreviousPerformRequestsWithTarget:self];
-}
-
-- (void) _startAllPeriodicOperations
-{
-    [self performSelector:@selector(updateMachineRecord) withObject:nil afterDelay:2];
-    [self performSelector:@selector(updateExperiment) withObject:nil afterDelay:3];
-    [self performSelector:@selector(updateRunInfo) withObject:nil afterDelay:3];
-    [self performSelector:@selector(updateDatabaseStats) withObject:nil afterDelay:4];
-    [self performSelector:@selector(periodicCompact) withObject:nil afterDelay:60];
 }
 
 #pragma mark ***Archival
